requirements.txt
langchain>=0.1.0
openai>=1.0.0
chromadb>=0.4.0
unstructured>=0.10.0
pypdf>=3.0.0
pyyaml>=6.0 